backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cloudinary==1.44.1
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
